requirements.txt
alembic==1.7.6
appdirs==1.4.4
APScheduler==3.8.1
beautifulsoup4==4.9.3
cachetools==5.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coloredlogs==14.3
docopt==0.6.2
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
fs==2.4.12
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
humanfriendly==9.1
idna==2.10
instalooter==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.6
MarkupSafe==1.1.1
multitasking==0.0.9
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
pdf2image==1.16.0
Pillow==8.0.1
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.24.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.1
SQLAlchemy==1.3.23
tenacity==6.3.1
tqdm==4.58.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.25.10
verboselogs==1.7
Werkzeug==1.0.1
yfinance==0.1.54
